--- src/main/groovy/ai/lamin/plugin/LaminFactory.groovy
@@ -1,36 +1,0 @@
-/*
- * Copyright 2025, Lamin Labs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ai.lamin.plugin
-
-import groovy.transform.CompileStatic
-import nextflow.Session
-import nextflow.trace.TraceObserver
-import nextflow.trace.TraceObserverFactory
-
-/**
- * Implements a factory object required to create
- * the {@link LaminObserver} instance.
- */
-@CompileStatic
-class LaminFactory implements TraceObserverFactory {
-
-    @Override
-    Collection<TraceObserver> create(Session session) {
-        return List.<TraceObserver>of(new LaminObserver())
-    }
-
-}
